refactor(persistence): improved GetAllStakedActors, moved from genesis.go
to actor.go
