+++ persistence/genesis.go
@@ -205,146 +205,3 @@ func (p PostgresContext) GetAllPools(height int64) (accs []modules.Account, err 
 	}
 	return
 }
-
-func (p PostgresContext) GetAllApps(height int64) (apps []modules.Actor, err error) {
-	ctx, tx, err := p.getCtxAndTx()
-	if err != nil {
-		return nil, err
-	}
-	rows, err := tx.Query(ctx, types.ApplicationActor.GetAllQuery(height))
-	if err != nil {
-		return nil, err
-	}
-	var actors []*types.Actor
-	for rows.Next() {
-		var actor *types.Actor
-		actor, height, err = p.getActorFromRow(rows)
-		if err != nil {
-			return
-		}
-		actors = append(actors, actor)
-	}
-	rows.Close()
-	for _, actor := range actors {
-		actorWithChains, err := p.getChainsForActor(ctx, tx, types.ApplicationActor, actor, height)
-		if err != nil {
-			return nil, err
-		}
-		apps = append(apps, actorWithChains)
-	}
-	return
-}
-
-func (p PostgresContext) GetAllValidators(height int64) (vals []modules.Actor, err error) {
-	ctx, tx, err := p.getCtxAndTx()
-	if err != nil {
-		return nil, err
-	}
-	rows, err := tx.Query(ctx, types.ValidatorActor.GetAllQuery(height))
-	if err != nil {
-		return nil, err
-	}
-	var actors []*types.Actor
-	for rows.Next() {
-		var actor *types.Actor
-		actor, height, err = p.getActorFromRow(rows)
-		if err != nil {
-			return
-		}
-		actors = append(actors, actor)
-	}
-	rows.Close()
-	for _, actor := range actors {
-		actor, err = p.getChainsForActor(ctx, tx, types.ApplicationActor, actor, height)
-		if err != nil {
-			return
-		}
-		vals = append(vals, actor)
-	}
-	return
-}
-
-func (p PostgresContext) GetAllServiceNodes(height int64) (sn []modules.Actor, err error) {
-	ctx, tx, err := p.getCtxAndTx()
-	if err != nil {
-		return nil, err
-	}
-	rows, err := tx.Query(ctx, types.ServiceNodeActor.GetAllQuery(height))
-	if err != nil {
-		return nil, err
-	}
-	var actors []*types.Actor
-	for rows.Next() {
-		var actor *types.Actor
-		actor, height, err = p.getActorFromRow(rows)
-		if err != nil {
-			return
-		}
-		actors = append(actors, actor)
-	}
-	rows.Close()
-	for _, actor := range actors {
-		actor, err = p.getChainsForActor(ctx, tx, types.ServiceNodeActor, actor, height)
-		if err != nil {
-			return
-		}
-		sn = append(sn, actor)
-	}
-	return
-}
-
-func (p PostgresContext) GetAllFishermen(height int64) (f []modules.Actor, err error) {
-	ctx, tx, err := p.getCtxAndTx()
-	if err != nil {
-		return nil, err
-	}
-	rows, err := tx.Query(ctx, types.FishermanActor.GetAllQuery(height))
-	if err != nil {
-		return nil, err
-	}
-	var actors []*types.Actor
-	for rows.Next() {
-		var actor *types.Actor
-		actor, height, err = p.getActorFromRow(rows)
-		if err != nil {
-			return
-		}
-		actors = append(actors, actor)
-	}
-	rows.Close()
-	for _, actor := range actors {
-		actor, err = p.getChainsForActor(ctx, tx, types.FishermanActor, actor, height)
-		if err != nil {
-			return
-		}
-		f = append(f, actor)
-	}
-	return
-}
-
-// IMPROVE: This is a proof of concept. Ideally we should have a single query that returns all actors.
-func (p PostgresContext) GetAllStakedActors(height int64) (actors []modules.Actor, err error) {
-	actors = make([]modules.Actor, 0)
-	var allActors []modules.Actor
-	allActors, err = p.GetAllValidators(height)
-	if err != nil {
-		return
-	}
-	actors = append(actors, allActors...)
-	allActors, err = p.GetAllServiceNodes(height)
-	if err != nil {
-		return
-	}
-	actors = append(actors, allActors...)
-	allActors, err = p.GetAllFishermen(height)
-	if err != nil {
-		return
-	}
-	actors = append(actors, allActors...)
-	allActors, err = p.GetAllApps(height)
-	if err != nil {
-		return
-	}
-	actors = append(actors, allActors...)
-	return
-}
